services/threads-scraper/pyproject.toml
[project]
name = "threads-scraper"
version = "1.0.0"
description = "使用 Playwright 抓取 Meta Threads 公開貼文的工具，支援排程、通知、用戶發現和 REST API"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [
    { name = "Vincent Xu", email = "vincentxuwork@gmail.com" }
]
keywords = ["threads", "scraper", "playwright", "webhook", "fastapi"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "fastapi>=0.133.1",
    "jmespath>=1.0.0",
    "nested-lookup>=0.2.25",
    "parsel>=1.8.0",
    "playwright>=1.40.0",
    "psycopg2-binary>=2.9.11",
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
    "schedule>=1.2.2",
    "uvicorn>=0.41.0",
]

[project.urls]
Homepage = "https://github.com/vincentxuwork/threads-scraper"
Repository = "https://github.com/vincentxuwork/threads-scraper"
Issues = "https://github.com/vincentxuwork/threads-scraper/issues"

[project.scripts]
threads-scheduler = "run_scheduler:main"
threads-api = "run_api:main"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
